Update customaction proj to use components in new location and rename msgraph to new namespace
