+++ packages/Calendar/Microsoft.BotFramework.CustomAction.CalendarSkill.csproj
@@ -26,7 +26,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\MsGraph\Microsoft.BotFramework.CustomAction.MsGraph.csproj" PrivateAssets="all" />
+    <ProjectReference Include="..\MsGraph\Microsoft.Bot.Component.MsGraph.csproj" PrivateAssets="all" />
   </ItemGroup>
 
   <ItemGroup>
