+++ packages/MsGraph/Actions/FindMeetingTimes.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Microsoft.BotFramework.Composer.CustomAction.Actions.MSGraph
+namespace Microsoft.Bot.Component.MsGraph.Actions.MSGraph
 {
     using System;
     using System.Collections.Generic;
@@ -11,7 +11,7 @@ namespace Microsoft.BotFramework.Composer.CustomAction.Actions.MSGraph
     using System.Threading.Tasks;
     using AdaptiveExpressions.Properties;
     using Microsoft.Bot.Builder.Dialogs;
-    using Microsoft.BotFramework.Composer.CustomAction.Models;
+    using Microsoft.Bot.Component.MsGraph.Models;
     using Microsoft.Graph;
     using Newtonsoft.Json;
