+++ packages/MsGraph/Actions/GetProfile.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Microsoft.BotFramework.Composer.CustomAction.Actions.MSGraph
+namespace Microsoft.Bot.Component.MsGraph.Actions.MSGraph
 {
     using System.Runtime.CompilerServices;
     using System.Threading;
