+++ packages/MsGraph/CalendarSkillTimeSlotModel.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Microsoft.BotFramework.Composer.CustomAction.Models
+namespace Microsoft.Bot.Component.MsGraph.Models
 {
     using System;
