+++ packages/MsGraph/MSGraphClient.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-namespace Microsoft.BotFramework.Composer.CustomAction
+namespace Microsoft.Bot.Component.MsGraph
 {
     using System;
     using System.Net.Http;
