+++ packages/MsGraph/Microsoft.Bot.Component.MsGraph.csproj
