+++ skills/declarative/calendarSkill/runtime/customaction/Microsoft.BotFramework.Composer.CustomAction.csproj
@@ -14,8 +14,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\..\..\experimental\components\MsGraph\Microsoft.BotFramework.CustomAction.MsGraph.csproj" />
-    <ProjectReference Include="..\..\..\..\..\experimental\components\Calendar\Microsoft.BotFramework.CustomAction.CalendarSkill.csproj" />
+    <ProjectReference Include="..\..\..\..\..\packages\Calendar\Microsoft.BotFramework.CustomAction.CalendarSkill.csproj" />
+    <ProjectReference Include="..\..\..\..\..\packages\MsGraph\Microsoft.Bot.Component.MsGraph.csproj" />
   </ItemGroup>
 
   <ItemGroup>
